requirements.txt
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
rich==13.7.1
aiolimiter==1.1.0
playwright==1.47.0
pandas==2.2.3
streamlit==1.40.2
plotly==5.24.1
